fix(search): use OS temp dir if no indexPath is set

+++ runtime/index.go
@@ -133,9 +133,11 @@ initialization:
 			case <-ctx.Done():
 				close(s.queue)
 
-				indexPath := getSearchIndexPath(s.cfg)
-				if indexPath != "" {
-					_ = os.RemoveAll(indexPath)
+				if !s.cfg.InMemory {
+					indexPath := getSearchIndexPath(s.cfg)
+					if indexPath != "" {
+						_ = os.RemoveAll(indexPath)
+					}
 				}
 
 				return
@@ -359,5 +361,5 @@ func getSearchIndexPath(cfg static.Search) string {
 	if indexPath == "" {
 		indexPath = os.TempDir()
 	}
-	return filepath.Join(cfg.IndexPath, "mokapi-bleve-index")
+	return filepath.Join(indexPath, "mokapi-bleve-index")
 }
